Removes kotest from docs, renames listener, add extension function for creating listener from container
