+++ modules/kotest/src/main/kotlin/org/testcontainers/containers/Extensions.kt
@@ -1,0 +1,6 @@
+package org.testcontainers.containers
+
+import org.testcontainers.lifecycle.Startable
+
+fun Startable.perTest() = StartablePerTestListener(this)
+fun Startable.perSpec() = StartablePerSpecListener(this)
