+++ modules/kotest/src/main/kotlin/org/testcontainers/containers/StartablePerSpecListener.kt
@@ -2,28 +2,29 @@ package org.testcontainers.containers
 
 import io.kotest.core.listeners.TestListener
 import io.kotest.core.spec.Spec
+import org.testcontainers.lifecycle.Startable
 
 
 /**
- * [ContainerPerSpecListener] starts the given [genericContainers] before execution of any test in the spec
+ * [StartablePerSpecListener] starts the given [startables] before execution of any test in the spec
  * and stops after execution of all tests.
  *
- * [genericContainers] can any of [GenericContainer] [ContainerisedDockerCompose] [LocalStackContainer] etc.
+ * [startables] can any of [GenericContainer] [ContainerisedDockerCompose] [LocalStackContainer] etc.
  *
  * This should be use when you want to a single container for all test in a single test class.
  *
  * @see
- * [ContainerPerTestListener]
+ * [StartablePerTestListener]
  * */
 
-class ContainerPerSpecListener(private vararg val genericContainers: GenericContainer<Nothing>) : TestListener {
+class StartablePerSpecListener(private vararg val startables: Startable) : TestListener {
     override suspend fun beforeSpec(spec: Spec) {
-        genericContainers.forEach { it.start() }
+        startables.forEach { it.start() }
         super.beforeSpec(spec)
     }
 
     override suspend fun afterSpec(spec: Spec) {
-        genericContainers.forEach { it.stop() }
+        startables.forEach { it.stop() }
         super.afterSpec(spec)
     }
 }
