+++ modules/kotest/src/main/kotlin/org/testcontainers/containers/StartablePerTestListener.kt
@@ -3,27 +3,28 @@ package org.testcontainers.containers
 import io.kotest.core.listeners.TestListener
 import io.kotest.core.test.TestCase
 import io.kotest.core.test.TestResult
+import org.testcontainers.lifecycle.Startable
 
 
 /**
- * [ContainerPerTestListener] starts the given [genericContainers] before execution of each test in the spec
+ * [StartablePerTestListener] starts the given [startables] before execution of each test in the spec
  * and stops after execution of each test.
  *
- * [genericContainers] can any of [GenericContainer] [ContainerisedDockerCompose] [LocalStackContainer] etc.
+ * [startables] can any of [GenericContainer] [ContainerisedDockerCompose] [LocalStackContainer] etc.
  *
  * This should be use when you want fresh container for each test.
  *
- * @see[ContainerPerSpecListener]
+ * @see[StartablePerSpecListener]
  * */
-class ContainerPerTestListener(private vararg val genericContainers: GenericContainer<Nothing>) : TestListener {
+class StartablePerTestListener(private vararg val startables: Startable) : TestListener {
 
    override suspend fun beforeTest(testCase: TestCase) {
-      genericContainers.forEach { it.start() }
+      startables.forEach { it.start() }
       super.beforeTest(testCase)
    }
 
    override suspend fun afterTest(testCase: TestCase, result: TestResult) {
-       genericContainers.forEach { it.stop() }
+       startables.forEach { it.stop() }
       super.afterTest(testCase, result)
    }
 }
