+++ modules/kotest/src/test/kotlin/org/testcontainers/containers/TestcontainerKotestIntegrationPerSpec.kt
@@ -1,28 +1,24 @@
-package com.example.kotest
+package org.testcontainers.containers
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-import kotlinx.coroutines.delay
-import org.testcontainers.containers.ContainerPerTestListener
-import org.testcontainers.containers.GenericContainer
 
-class TestContainerKotestIntegrationPerTest : StringSpec({
+class TestcontainerKotestIntegrationPerSpec : StringSpec({
     val container = GenericContainer<Nothing>("ubuntu")
     setContainerCommand(container) // Set container command for first time to echo hello Kotest
 
     // Register the Listener
-    listeners(ContainerPerTestListener(container))
+    listeners(container.perSpec())
 
     "test should echo hello Kotest for first test" {
         container.logs.trim() shouldBe "hello Kotest"
         // Set container command so that next time when container start we will have hello Kotest again in container log
         setContainerCommand(container)
-        delay(1000)
     }
 
     "test should echo hello Kotest for second spec as well" {
-        // presence of hello Kotest again verifies that container is restarted.
-        container.logs.trim() shouldBe "hello Kotest again"
+        // presence of "hello Kotest" verifies that container is not restarted.
+        container.logs.trim() shouldBe "hello Kotest"
     }
 
 })
