+++ settings.gradle
@@ -54,7 +54,6 @@ include ':docs:examples:junit4:generic'
 include ':docs:examples:junit4:redis'
 include ':docs:examples:junit5:redis'
 include ':docs:examples:spock:redis'
-include ':docs:examples:kotest'
 
 include 'test-support'
